Translator を抽象クラスに変更
